+++ exonum/src/sandbox/consensus/timeouts.rs
@@ -19,7 +19,7 @@ use std::time::Duration;
 use crate::crypto::CryptoHash;
 use crate::helpers::{Height, Round, ValidatorId};
 use crate::node::state::PROPOSE_REQUEST_TIMEOUT;
-use crate::sandbox::{sandbox::timestamping_sandbox, sandbox_tests_helper::*};
+use crate::sandbox::{sandbox_tests_helper::*, timestamping_sandbox};
 
 /// HANDLE ROUND TIMEOUT:
 /// - Ignore if height and round are not the same
