+++ exonum/src/sandbox/sandbox_tests_helper.rs
@@ -18,7 +18,7 @@ use bit_vec::BitVec;
 use std::{cell::RefCell, collections::BTreeMap, time::Duration};
 
 use super::timestamping::DATA_SIZE;
-use super::{sandbox::Sandbox, timestamping::TimestampingTxGenerator};
+use super::{timestamping::TimestampingTxGenerator, Sandbox};
 use crate::blockchain::Block;
 use crate::crypto::{CryptoHash, Hash, HASH_SIZE};
 use crate::helpers::{Height, Milliseconds, Round, ValidatorId};
@@ -424,7 +424,7 @@ pub fn add_one_height_with_transactions_from_other_validator(
 ) -> Vec<Hash> {
     // sort transaction in order accordingly their hashes
     let mut tx_pool = BTreeMap::new();
-    tx_pool.extend(txs.into_iter().map(|tx| (tx.hash(), tx.clone())));
+    tx_pool.extend(txs.iter().map(|tx| (tx.hash(), tx.clone())));
     let raw_txs = tx_pool.values().cloned().collect::<Vec<_>>();
     let txs: &[Signed<RawTransaction>] = raw_txs.as_ref();
