+++ exonum/src/storage/keys.rs
@@ -502,11 +502,11 @@ mod tests {
         let (mut buffer1, mut buffer2) = ([0_u8; 12], [0_u8; 12]);
         for _ in 0..FUZZ_SAMPLES {
             let time1 = Utc.timestamp(
-                rng.gen::<i64>() % (i32::max_value() as i64),
+                rng.gen::<i64>() % i64::from(i32::max_value()),
                 rng.gen::<u32>() % 1_000_000_000,
             );
             let time2 = Utc.timestamp(
-                rng.gen::<i64>() % (i32::max_value() as i64),
+                rng.gen::<i64>() % i64::from(i32::max_value()),
                 rng.gen::<u32>() % 1_000_000_000,
             );
             time1.write(&mut buffer1);
@@ -619,8 +619,8 @@ mod tests {
     fn decimal_round_trip() {
         let decimals = [
             Decimal::from_str("3.14").unwrap(),
-            Decimal::from_parts(1102470952, 185874565, 1703060790, false, 28),
-            Decimal::new(9497628354687268, 12),
+            Decimal::from_parts(1_102_470_952, 185_874_565, 1_703_060_790, false, 28),
+            Decimal::new(9_497_628_354_687_268, 12),
             Decimal::from_str("0").unwrap(),
             Decimal::from_str("-0.000000000000000000019").unwrap(),
         ];
